+++ semana19/LaBook/.gitignore
@@ -1,0 +1,5 @@
+node_modules
+build
+.env
+.vscode
+package-lock.json
