+++ semana19/LaBook/src/business/services/authenticationData.ts
@@ -1,0 +1,27 @@
+import { authenticationData } from "../entities/user"
+import * as jwt from "jsonwebtoken"
+
+export function generateToken(
+    payload: authenticationData
+ ): string {
+    return jwt.sign(
+       payload,
+       process.env.JWT_KEY as string,
+       {
+          expiresIn: process.env.JWT_EXPIRES_IN
+       }
+    )
+ }
+ 
+ export function getTokenData(
+    token: string
+ ): authenticationData {
+    const result: any = jwt.verify(
+       token,
+       process.env.JWT_KEY as string
+    )
+ 
+    return { id: result.id, }
+ }
+ 
+
