+++ semana19/LaBook/src/business/services/hashManager.ts
@@ -1,0 +1,16 @@
+import * as bcrypt from 'bcryptjs';
+
+export const hash = async (
+    plainText: string
+ ): Promise<string> => {
+    const rounds = Number(process.env.BCRYPT_COST);
+    const salt = await bcrypt.genSalt(rounds);
+    return bcrypt.hash(plainText, salt)
+ }
+ 
+ const compare = async (
+    plainText: string, cypherText: string
+ ): Promise<boolean> => {
+    return bcrypt.compare(plainText, cypherText)
+ }
+
