+++ semana19/LaBook/src/data/model/connection.ts
@@ -1,0 +1,16 @@
+import Knex from "knex"
+import dotenv from "dotenv"
+import knex from "knex"
+
+dotenv.config()
+
+export const connection: Knex = knex({
+   client: "mysql",
+   connection: {
+      host: process.env.DB_HOST,
+      user: process.env.DB_USER,
+      password: process.env.DB_PASSWORD,
+      port: 3306,
+      database: process.env.DB_NAME,
+   }
+})
