+++ semana19/LaBook/src/index.ts
@@ -2,27 +2,8 @@
 
 import express, { Express, Request, Response } from "express"
 import cors from "cors"
-import knex from "knex"
-import dotenv from "dotenv"
-import * as jwt from "jsonwebtoken"
-import * as bcrypt from "bcryptjs"
-import { v4 } from "uuid"
-import Knex from "knex"
 
 /**************************** CONFIG ******************************/
-
-dotenv.config()
-
-export const connection: Knex = knex({
-   client: "mysql",
-   connection: {
-      host: process.env.DB_HOST,
-      user: process.env.DB_USER,
-      password: process.env.DB_PASSWORD,
-      port: 3306,
-      database: process.env.DB_NAME,
-   }
-})
 
 const app: Express = express()
 app.use(express.json())
@@ -30,71 +11,13 @@ app.use(cors())
 
 /**************************** TYPES ******************************/
 
-type AuthenticationData = {
-   id: string
-}
 
-type User = {
-   id: string,
-   name: string,
-   email: string,
-   password: string
-}
 
-enum POST_TYPES {
-   NORMAL = "normal",
-   EVENT = "event"
-}
 
-type Post = {
-   id: string,
-   photo: string,
-   description: string,
-   type: POST_TYPES,
-   createdAt: Date,
-   authorId: string
-}
 
 /**************************** SERVICES ******************************/
 
-const generateId = (): string => v4()
 
-function generateToken(
-   payload: AuthenticationData
-): string {
-   return jwt.sign(
-      payload,
-      process.env.JWT_KEY as string,
-      {
-         expiresIn: process.env.JWT_EXPIRES_IN
-      }
-   )
-}
-
-function getTokenData(
-   token: string
-): AuthenticationData {
-   const result: any = jwt.verify(
-      token,
-      process.env.JWT_KEY as string
-   )
-
-   return { id: result.id, }
-}
-
-const hash = async (
-   plainText: string
-): Promise<string> => {
-   const rounds = Number(process.env.BCRYPT_COST);
-   const salt = await bcrypt.genSalt(rounds);
-   return bcrypt.hash(plainText, salt)
-}
-
-const compare = async (
-   plainText: string, cypherText: string
-): Promise<boolean> => {
-   return bcrypt.compare(plainText, cypherText)
-}
 
 /**************************** ENDPOINTS ******************************/
 
@@ -113,13 +36,13 @@ app.post('/users/signup', async (req: Request, res: Response) => {
 
       const cypherPassword = await hash(password);
 
-      await connection('labook_users')
-         .insert({
-            id,
-            name,
-            email,
-            password: cypherPassword
-         })
+      // await connection('labook_users')
+      //    .insert({
+      //       id,
+      //       name,
+      //       email,
+      //       password: cypherPassword
+      //    })
 
       const token: string = generateToken({ id })
 
@@ -145,9 +68,9 @@ app.post('/users/login', async (req: Request, res: Response) => {
          throw new Error(message)
       }
 
-      const queryResult: any = await connection("labook_users")
-         .select("*")
-         .where({ email })
+      // const queryResult: any = await connection("labook_users")
+      //    .select("*")
+      //    .where({ email })
 
       if (!queryResult[0]) {
          res.statusCode = 401
@@ -155,7 +78,7 @@ app.post('/users/login', async (req: Request, res: Response) => {
          throw new Error(message)
       }
 
-      const user: User = {
+      const user: user = {
          id: queryResult[0].id,
          name: queryResult[0].name,
          email: queryResult[0].email,
@@ -196,14 +119,14 @@ app.post('/posts/create', async (req: Request, res: Response) => {
 
       const id: string = generateId()
 
-      await connection("labook_posts")
-         .insert({
-            id,
-            photo,
-            description,
-            type,
-            author_id: tokenData.id
-         })
+      // await connection("labook_posts")
+      //    .insert({
+      //       id,
+      //       photo,
+      //       description,
+      //       type,
+      //       author_id: tokenData.id
+      //    })
 
       res.status(201).send({ message })
 
@@ -221,9 +144,9 @@ app.get('/posts/:id', async (req: Request, res: Response) => {
 
       const { id } = req.params
 
-      const queryResult: any = await connection("labook_posts")
-         .select("*")
-         .where({ id })
+      // const queryResult: any = await connection("labook_posts")
+      //    .select("*")
+      //    .where({ id })
 
       if (!queryResult[0]) {
          res.statusCode = 404
